refactor: Index Kurtosis packages by their github location instead of their name (#36)
This is because right now nothing prevents package authors from setting
the name in kurtosis.yml to something non unique.

This PR also updates how packages are updated in the crawling logic. We
now first iterate through the list of currently stored package to update
them (since we know their github coordinate already), and only after we
search for new packages on Github. This has the advantage of being more
robust to random behaviour Github searches might have on the long term