feat: generate typescript bindings (#46)
For the new enclave manager work we should use the well typed client
that could be generated for us from the indexer spec.

This PR adds support for generating typescript bindings for this
interface so that a well typed client can be generated by consumers.

Ideally this typescript package would be built and published to an npm
repo, where it could then be pulled back into the enclave manager. For
now though I'm going to just copy the build output into the enclave
manager (this badness seems consistent with how we obtain the
enclave-manager-sdk there).